Add files via upload
The goal of this lab is to experiment with accessing serial port directly without using
Arduino IDE, but using a C program instead. You will be reading data collected from a
ultrasonic sensor that measures the distance between an object and itself (aka
rangefinding). With the data collected, you will be constructing a Binary Search Tree using
the collected measurements.